refactor( Mapper.kt ): move it to the service layer, add new functions, format with ktlint
