+++ src/main/kotlin/ecommerce/service/CartService.kt
@@ -5,7 +5,7 @@ import ecommerce.dto.CartItemResponse
 import ecommerce.exception.NotFoundException
 import ecommerce.model.Cart
 import ecommerce.model.CartItem
-import ecommerce.model.mapper.CartItemMapper
+import ecommerce.service.mapper.CartItemMapper
 import ecommerce.repository.CartRepository
 import ecommerce.repository.MemberRepository
 import ecommerce.repository.ProductRepository
