+++ src/main/kotlin/ecommerce/service/mapper/Mapper.kt
@@ -1,0 +1,68 @@
+package ecommerce.service.mapper
+
+import ecommerce.dto.CartItemResponse
+import ecommerce.dto.OptionResponse
+import ecommerce.dto.PagedResponse
+import ecommerce.dto.ProductResponse
+import ecommerce.dto.RegisteredMember
+import ecommerce.dto.Role
+import ecommerce.model.CartItem
+import ecommerce.model.Member
+import ecommerce.model.Option
+import ecommerce.model.Product
+import org.springframework.data.domain.PageImpl
+
+object CartItemMapper {
+    fun toResponse(cartItem: CartItem) =
+        CartItemResponse(
+            quantity = cartItem.quantity,
+            productId = cartItem.product.id ?: 0,
+            productName = cartItem.product.name,
+            productPrice = cartItem.product.price,
+            productImageUrl = cartItem.product.imageUrl,
+        )
+
+    fun toPagedResponse(productPage: PageImpl<CartItemResponse>) =
+        PagedResponse(
+            productPage.content,
+            productPage.number,
+            productPage.size,
+            productPage.totalPages,
+            productPage.totalElements,
+        )
+}
+
+object OptionMapper {
+    fun toOptionResponse(option: Option) =
+        OptionResponse(
+            id = option.id,
+            name = option.name,
+            quantity = option.quantity,
+        )
+}
+
+object ProductMapper {
+    fun toProductResponse(entity: Product) =
+        ProductResponse(
+            id = entity.id,
+            name = entity.name,
+            price = entity.price,
+            imageUrl = entity.imageUrl,
+        )
+
+    fun toPagedResponse(productPage: PageImpl<ProductResponse>) =
+        PagedResponse(
+            productPage.content,
+            productPage.number,
+            productPage.size,
+            productPage.totalPages,
+            productPage.totalElements,
+        )
+}
+
+object MemberMapper {
+    fun toRegisteredMember(member: Member): RegisteredMember {
+        val role = Role.valueOf(member.role)
+        return RegisteredMember(member.id!!, member.email, role)
+    }
+}
