+++ src/test/kotlin/ecommerce/controller/ProductControllerMockTest.kt
@@ -7,7 +7,7 @@ import ecommerce.dto.RegisteredMember
 import ecommerce.dto.Role
 import ecommerce.infrastructure.AuthorizationExtractor
 import ecommerce.infrastructure.JwtTokenProvider
-import ecommerce.model.mapper.ProductMapper
+import ecommerce.service.mapper.ProductMapper
 import ecommerce.service.ProductService
 import org.assertj.core.api.Assertions.assertThat
 import org.json.JSONObject
@@ -79,7 +79,7 @@ class ProductControllerMockTest
         fun `test getPages`() {
             setMockMemberTo(Role.USER)
 
-            val testProduct = ProductMapper.toProductDto(DatabaseFixture.createPaintingSadHuman())
+            val testProduct = ProductMapper.toProductResponse(DatabaseFixture.createPaintingSadHuman())
             whenever(productService.getPages(any(), any()))
                 .thenReturn(PageImpl(listOf(testProduct), PageRequest.of(0, 10), 20))
